+++ backend/booking-service/src/main/java/com/hotel/booking/payment/PlatnoscRepository.java
@@ -2,5 +2,5 @@ package com.hotel.booking.payment;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PlatnoscRepository extends JpaRepository<Platnosc, Long> {
+public interface PlatnoscRepository extends JpaRepository<Platnosc, Integer> {
 }
