+++ backend/booking-service/src/main/java/com/hotel/booking/reservation/RezerwacjaRepository.java
@@ -2,5 +2,4 @@ package com.hotel.booking.reservation;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RezerwacjaRepository extends JpaRepository<Rezerwacja, Long> {
-}
+public interface RezerwacjaRepository extends JpaRepository<Rezerwacja, Integer> {}
