+++ backend/catalog-service/src/main/java/com/hotel/catalog/meal/WyzywienieRepository.java
@@ -2,5 +2,5 @@ package com.hotel.catalog.meal;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WyzywienieRepository extends JpaRepository<Wyzywienie, Long> {
+public interface WyzywienieRepository extends JpaRepository<Wyzywienie, Integer> {
 }
